+++ lib/shoulda/matchers/active_record/association_matchers/dependent_matcher.rb
@@ -19,10 +19,10 @@ module Shoulda
           def matches?(subject)
             self.subject = ModelReflector.new(subject, name)
 
-            if option_verifier.correct_for_string?(:dependent, dependent)
+            if option_matches?
               true
             else
-              self.missing_option = "#{name} should have #{dependent} dependency"
+              self.missing_option = generate_missing_option
               false
             end
           end
@@ -31,8 +31,29 @@ module Shoulda
 
           attr_accessor :subject, :dependent, :name
 
+          private
+
           def option_verifier
             @option_verifier ||= OptionVerifier.new(subject)
+          end
+
+          def option_matches?
+            option_verifier.correct_for?(option_type, :dependent, dependent)
+          end
+
+          def option_type
+            case dependent
+              when true, false then :boolean
+              else :string
+            end
+          end
+
+          def generate_missing_option
+            [
+              "#{name} should have",
+              (dependent == true ? 'a' : dependent),
+              'dependency'
+            ].join(' ')
           end
         end
       end